Align runtime scaffolding with Ruff type-hint rules
