+++ src/diaremot/pipeline/diarization/__init__.py
@@ -1,0 +1,24 @@
+from __future__ import annotations
+
+from .clustering import SpectralClusterer, build_agglo
+from .config import DiarizationConfig, DiarizedTurn
+from .embeddings import ECAPAEncoder
+from .paths import MODEL_ROOTS, iter_model_subpaths
+from .pipeline import SpeakerDiarizer
+from .registry import SpeakerRegistry
+from .segments import collapse_single_speaker_turns
+from .vad import SileroVAD
+
+__all__ = [
+    "DiarizationConfig",
+    "DiarizedTurn",
+    "SpeakerDiarizer",
+    "SpeakerRegistry",
+    "ECAPAEncoder",
+    "SileroVAD",
+    "SpectralClusterer",
+    "build_agglo",
+    "collapse_single_speaker_turns",
+    "MODEL_ROOTS",
+    "iter_model_subpaths",
+]
