+++ src/diaremot/pipeline/runtime/__init__.py
@@ -1,0 +1,15 @@
+"""Pipeline runtime scaffolding aligned with the paralinguistics package."""
+
+from .environment import PipelineEnvironment, bootstrap_environment
+from .executor import StageExecutor, execute_plans
+from .session import PipelineSession, StagePlan, build_stage_plan
+
+__all__ = [
+    "PipelineEnvironment",
+    "PipelineSession",
+    "StagePlan",
+    "StageExecutor",
+    "bootstrap_environment",
+    "build_stage_plan",
+    "execute_plans",
+]
